Ver usuários (Teste)
atualização no front. adicionando funcionlidade de teste " botão para que o usuário logado acesse a lista de todos os usuários utilizando seu token e retornando o  conteudo html da resposta em uma div"

+++ projeto/src/main/java/br/com/projetoweb/projeto/service/UsuarioService.java
@@ -2,13 +2,15 @@ package br.com.projetoweb.projeto.service;
 
 import java.util.List;
 import java.util.Optional;
+
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.ui.Model;
-
 import br.com.projetoweb.projeto.Security.Token;
 import br.com.projetoweb.projeto.Security.TokenUtil;
 import br.com.projetoweb.projeto.dto.UsuarioCadastroDTO;
@@ -56,10 +58,14 @@ public class UsuarioService {
 		return usuarioEditado;
 	}
 	
-	public Boolean deletarUsuario(Integer id) {
-		repository.deleteById(id);
-		return true;
-		
+	public ResponseEntity <?> deletarUsuario(Integer id) {
+		Optional<Usuario> usuario = repository.findById(id);
+		if (usuario.isPresent()){
+			repository.deleteById(id);
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Usuário deletado com sucesso!");
+		}else{
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não econtrado");
+		}
 	}
 	
 	public ResponseEntity<?> loginUsuario(UsuarioLoginDTO usuario) {	
@@ -71,7 +77,10 @@ public class UsuarioService {
 					String tokenAuth = TokenUtil.createToken(loginUsuario.get().getEmail());
 					Token token = new Token(tokenAuth);
 					UsuarioLogadoDTO usuarioLogadoDTO = new UsuarioLogadoDTO(loginUsuario.get().getId(),loginUsuario.get().getEmail(),loginUsuario.get().getNome(),loginUsuario.get().getTelefone());
-					return ResponseEntity.ok().header("Authorization",token.getToken()).body(usuarioLogadoDTO);
+					JSONObject responseJson = new JSONObject(usuarioLogadoDTO);
+					responseJson.put("token", tokenAuth);
+					String responseBody = responseJson.toString();
+					return ResponseEntity.status(HttpStatus.OK).header(HttpHeaders.AUTHORIZATION,token.getToken()).body(responseBody);
 			}else{
 				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Email ou Senha incorretos!");
 			}
